backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
pydantic-settings
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pytest
flake8
